Create Dynamic Nesting  - React Typescript
I hate nested objects. So I did a bunch of exercise creating them. This one creates objects that are nested dynamically under inputted keys.

I released the structure of the nested object is similar to linked list. So I added couple of functionality of deleting a key(element) or inserting a key at a given point.